pyproject.toml
[project]
name = "dump"
version = "0.1.0"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pyinfra~=3.5.1",
    "azure-identity~=1.25.1",
    "azure-mgmt-resourcegraph~=8.0.1",
]

[build-system]
requires = ["uv_build~=0.9.7"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
namespace = true

[project.entry-points."pyinfra.connectors"]
azure = 'dump.connectors.microsoft_azure:AzureConnector'